refactor: OfficialScheduleTooltip 분리
